% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_error.R
\name{add_prop_noise}
\alias{add_prop_noise}
\title{add proportional noise}
\usage{
add_prop_noise(n_rep, value, sd, use_zero = FALSE)
}
\arguments{
\item{n_rep}{number of replicates}

\item{value}{the values to add error to}

\item{sd}{the noise}

\item{use_zero}{should values be returned with zeros}
}
\value{
matrix
}
\description{
add proportional noise
}

