% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_error.R
\name{add_prop_uniform}
\alias{add_prop_uniform}
\title{add additive and proportional noise}
\usage{
add_prop_uniform(n_rep, value, u_sd, p_sd)
}
\arguments{
\item{n_rep}{number of replicates}

\item{value}{the initial values}

\item{u_sd}{uniform or additive error}

\item{p_sd}{proportional error}
}
\value{
matrix
}
\description{
add additive and proportional noise
}

