% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_error.R
\name{add_uniform_noise}
\alias{add_uniform_noise}
\title{add uniform noise}
\usage{
add_uniform_noise(n_rep, value, sd, use_zero = FALSE)
}
\arguments{
\item{n_rep}{number of replicates}

\item{value}{the value to use}

\item{sd}{the error to add}

\item{use_zero}{should stuff be reported using zero (used when doing two types of noise)}
}
\value{
matrix
}
\description{
add uniform noise
}

